+++ src/web-services/interner-web-service/_controllers/controller.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const parameter_list = require('../_models/parameter_list')
+
+exports.get_parameter_list = function (req, res) {
+    res.json(parameter_list);
+}
