+++ src/web-services/interner-web-service/_models/parameter_list.json
@@ -1,0 +1,13 @@
+[{
+        "path": "/",
+        "description": "lol"
+    },
+    {
+        "path": "/api",
+        "description": "lol2"
+    },
+    {
+        "path": "/log",
+        "description": "lol3"
+    }
+]
