+++ src/web-services/interner-web-service/api/_models/parameter_list.json
@@ -1,0 +1,10 @@
+[
+    {
+        "path": "/",
+        "description": "lol"
+    },
+    {
+        "path": "/v1",
+        "description": "lol2"
+    }
+]
