+++ src/web-services/interner-web-service/api/v1/_controllers/controller_root.js
@@ -1,0 +1,5 @@
+'use strict';
+
+exports.root = function (req, res) {
+    res.send('v1');
+};
