--- src/web-services/interner-web-service/api/v1/models/database/database.json
@@ -1,48 +1,0 @@
-[{
-		"name": "Akropolis von Athen",
-		"koordinaten": {
-			"lon": 37.971532,
-			"lat": 23.725749
-		},
-		"bild": {
-			"link": "https://cdn.holidayguru.ch/wp-content/uploads/2017/06/urlaubsguru.de_acropolis-greece-parthenon-temple-istock_000021610339_large.jpg"
-		},
-		"beschreibung": "Die Akropolis in Athen ist die wohl bekannteste Vertreterin der als Akropolis bezeichneten Stadtfestungen des antiken Griechenlands. Die Athener Akropolis mit ihren bemerkenswerten Gebäuden wird deshalb oft einfach \"die Akropolis\" genannt. Den ältesten Teil der Stadt Athen ließ Perikles nach der Zerstörung durch die Perser unter Leitung des berühmten Bildhauers Phidias von den Architekten Iktinos und Kallikrates sowie Mnesikles neu bebauen. Auf einem flachen, 156 Meter hohen Felsen stehen die zwischen 467 v. Chr. und 406 v. Chr. erbauten Propyläen, das Erechtheion, der Niketempel und der Parthenon, in dem eine kolossale Statue der Göttin Athene aus Gold und Elfenbein stand. Eine stark verkleinerte Replik der Statue befindet sich im Archäologischen Nationalmuseum in Athen. Die Akropolis in Athen ist seit 1986 Teil des UNESCO-Weltkulturerbes."
-	},
-
-	{
-		"name": "Parthenon",
-		"koordinaten": {
-			"lon": 37.9715285,
-			"lat": 23.7267166
-		},
-		"beschreibung": "Der Parthenon (gr. Jungfrauengemach) ist der Tempel für die Stadtgöttin Pallas Athena Parthenos auf der Athener Akropolis. Er wurde zum Dank für die Rettung der Athener und Griechen durch die Göttin nach dem letzten Perserkrieg als dorischer Peripteros erbaut. Im Laufe der Geschichte Griechenlands diente das Gebäude unter anderem auch als Schatzkammer des Attischen Seebunds. Der Parthenon ist eines der berühmtesten noch existierenden Baudenkmäler des antiken Griechenlands und eines der bekanntesten Gebäude weltweit. Das Gebäude beherrscht als zentraler Bau seit fast 2.500 Jahren die Athener Akropolis."
-	},
-
-	{
-		"name": "Plaka",
-		"koordinaten": {
-			"lon": 37.972222,
-			"lat": 23.730556
-		},
-		"beschreibung": "Die Plaka ist einer der ältesten Stadtteile Athens am Fuß der Akropolis. Sie liegt zwischen den Metrostationen Akropolis und Monastiraki und grenzt an Psirri und die heutige Innenstadt. Der Stadtteil und seine Bausubstanz können auf eine Kontinuität bis in die Antike zurückblicken. Viele Häuser haben Grundmauern aus der Antike, wenn auch die Fassaden zumeist aus dem 18. Jahrhundert stammen. Bis zur modernen Stadtplanung des frühen 19. Jahrhunderts bildete sie den Kern der Stadt Athen. 1840 lebten noch 19.000 Menschen in der Altstadt. Danach wandelte sich die Plaka zum Tavernenviertel, das in allen gesellschaftlichen Schichten gleich beliebt war. Seit den 1990er Jahren begann ein Wiederaufleben des Stadtteils, viele alte Häuser wurden restauriert, nahezu der ganze Stadtteil verkehrsberuhigt. Die Immobilienpreise reichen mittlerweile an die teuersten Athener Stadtteile heran."
-	},
-
-	{
-		"name": "Olympieion",
-		"koordinaten": {
-			"lon": 37.969167,
-			"lat": 23.733611
-		},
-		"beschreibung": "Das Olympieion (auch Tempel des Olympischen Zeus) in Athen war einer der größten Tempel im antiken Griechenland. Der Bau geht auf das 6. Jahrhundert v. Chr. zurück, wurde aber erst unter dem römischen Kaiser Hadrian im 2. Jahrhundert n. Chr. vollendet. Das Olympieion befindet sich rund 500 m östlich der Akropolis. Die ältesten ausgegrabenen Baureste an dieser Stelle stammen wohl von einem Tempel aus der ersten Hälfte des 6. Jahrhunderts v. Chr., der wahrscheinlich unter der Tyrannis des Peisistratos erbaut wurde. Seine Söhne Hippias und Hipparchos ließen an dessen Stelle einen riesenhaften Tempel errichten, wohl nach dem Vorbild der kurz vorher begonnenen Dipteroi in Ionien. Nachdem Sulla (römischer Politiker und Diktator) 86 v.& Chr. Athen erobert hatte, ließ er zwei Säulen des Olympieion nach Rom bringen und im Tempel des Jupiter Capitolinus verbauen. Wann und wie genau das Olympieion zerstört wurde, ist unbekannt, wahrscheinlich jedoch durch ein Erdbeben im Mittelalter. Danach wurden die Ruinen als Steinbruch benutzt. Mit archäologischen Ausgrabungen wurde 1889 begonnen."
-	},
-
-	{
-		"name": "Syntagma-Platz",
-		"koordinaten": {
-			"lon": 37.975627,
-			"lat": 23.735484
-		},
-		"beschreibung": "Der Syntagma-Platz (= „Platz der Verfassung“) ist ein Platz im Zentrum Athens. 1837 wurde der Platz als Musenplatz angelegt und war das Zentrum der neuen Innenstadt. Infolge der Revolution des 3. September 1843 verkündete König Otto von einem Balkon des königlichen Schlosses seine Zustimmung zur konstitutionellen Monarchie. So erhielt der Syntagma-Platz seinen heutigen Namen. Während des 20. Jahrhunderts war er von Kaffeehäusern in den umliegenden Gebäuden geprägt, die auf den Platz Tische und Stühle stellten. Während des U-Bahn-Baus in den 1990er-Jahren wurden Fundamente von Gebäuden aus der Antike entdeckt, wenngleich dies in Athen keine Seltenheit darstellt, entschlossen sich die Planer diese in den Bau der Station zu integrieren, ein Teil ist sichtbar hinter einer Verglasung an der Ecke des Platzes, ein weiterer Teil an dem 2003 eröffneten Knotenpunkt „Syntagma“ der Metro Athen."
-	}
-]
