+++ src/web-services/interner-web-service/log/_models/error.log
@@ -1,0 +1,16 @@
+{ message:
+   '500 - Cannot set headers after they are sent to the client - / - GET - ::ffff:127.0.0.1',
+  level: 'error',
+  timestamp: '2019-04-04T14:53:20.105Z' }
+{ message:
+   '500 - Cannot set headers after they are sent to the client - / - GET - ::ffff:127.0.0.1',
+  level: 'error',
+  timestamp: '2019-04-04T14:54:00.663Z' }
+{ message:
+   '500 - Cannot set headers after they are sent to the client - / - GET - ::ffff:127.0.0.1',
+  level: 'error',
+  timestamp: '2019-04-04T14:54:29.751Z' }
+{ message:
+   '500 - Cannot set headers after they are sent to the client - /api - GET - ::ffff:127.0.0.1',
+  level: 'error',
+  timestamp: '2019-04-04T14:55:23.923Z' }
