+++ src/web-services/interner-web-service/log/_models/info.log
@@ -1,0 +1,27 @@
+{ message: '/log/info - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:03:32.952Z' }
+{ message: '/log/info - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:03:35.289Z' }
+{ message: '/log/error - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:03:39.361Z' }
+{ message: '/log/error - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:05:06.374Z' }
+{ message: '/api - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:13:48.090Z' }
+{ message: '/api - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:19:10.075Z' }
+{ message: '/ - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:19:12.462Z' }
+{ message: '/log - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:19:33.297Z' }
+{ message: '/log/error - GET - ::1',
+  level: 'info',
+  timestamp: '2019-04-04T20:19:36.753Z' }
