+++ src/web-services/interner-web-service/log/log.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const router = require('express').Router();
+
+const controller = require('./_controllers/controller');
+
+router.get('/', controller.root);
+router.get('/info', controller.get_log_info);
+router.get('/error', controller.get_log_error);
+
+module.exports = router;
